feat: add data explorer URL support in dataset configurations and enhance URL resolution logic
