+++ statgpt/common/data/base/config.py
@@ -198,10 +198,23 @@ DIMENSION_CONFIG_TYPES = Annotated[
 class BaseDataSetConfig(BaseModel):
     is_official: bool = Field(default=False)
     citation: DatasetCitation | None = Field(default=None)
+    data_explorer_url: str | None = Field(
+        default=None,
+        description=(
+            "Base URL of the data explorer UI for this dataset. When set, it overrides the "
+            "data source default for explorer links in query results and for dataset_url when "
+            "no citation URL is configured."
+        ),
+    )
     indexer: Annotated[IndexerConfig | None, IndexingField()] = Field(default=None)
     pinned_columns: list[str] = Field(
         description="Column names and order to pin in the data in grid", default_factory=list
     )
+
+    def get_data_explorer_url(self) -> str | None:
+        if self.data_explorer_url:
+            return replace_env(self.data_explorer_url)
+        return None
 
 
 class DataSetConfigTemplate(BaseDataSetConfig):
