+++ statgpt/common/data/quanthub/v21/dataset.py
@@ -3,15 +3,12 @@ import typing
 import uuid
 from collections.abc import Iterable
 
-import pandas as pd
 from sdmx.message import StructureMessage
 from sdmx.model.common import BaseAnnotation
 from sdmx.model.v21 import ContentConstraint
 from sdmx.model.v21 import DataflowDefinition as DataFlow
-from sdmx.model.v21 import DataStructureDefinition
 
-from statgpt.common.auth.auth_context import AuthContext
-from statgpt.common.data.base import DataResponseStatus, DataSetAvailabilityQuery, DataSetQuery
+from statgpt.common.data.base import DataSetAvailabilityQuery
 from statgpt.common.data.base.property_source import PropertySource, PropertySourceEnum
 from statgpt.common.data.base.sdmx_schemas import Sdmx30AnnotationModel
 from statgpt.common.data.base.updated_at_mixin import UpdatedAtMixin
@@ -19,10 +16,6 @@ from statgpt.common.data.quanthub.config import QuanthubDataSetConfig
 from statgpt.common.data.sdmx import Sdmx21DataSet
 from statgpt.common.data.sdmx.common import SdmxDimension
 from statgpt.common.data.sdmx.v21.attribute import Sdmx21Attribute
-from statgpt.common.data.sdmx.v21.dataset import Sdmx21DataResponse
-from statgpt.common.data.sdmx.v21.query import SdmxDataSetQuery, SdmxQueryReadinessStatus
-from statgpt.common.data.sdmx.v21.utils import convert_keys_to_str
-from statgpt.common.schemas.enums import DataParsingStatus, DataRequestStatus
 
 if typing.TYPE_CHECKING:
     from statgpt.common.data.quanthub.v21.datasource import QuanthubSdmx21DataSourceHandler
@@ -62,10 +55,13 @@ class QuanthubSdmx21DataSet(UpdatedAtMixin, Sdmx21DataSet):
     @property
     def dataset_url(self) -> str | None:
         if self._datasource.config.use_data_explorer_for_dataset_url:  # type: ignore
-            if data_explorer_url := self._datasource.config.get_data_explorer_url():  # type: ignore
+            if data_explorer_url := self._resolved_data_explorer_base_url():
                 return f"{data_explorer_url}?urn={self._short_urn}"
             else:
-                _log.warning("Data explorer URL is not configured for the data source: %s", self._datasource.source_id)  # type: ignore
+                _log.warning(
+                    "Data explorer URL is not configured on the dataset or data source: %s",
+                    self._datasource.source_id,  # type: ignore
+                )
         return super().dataset_url
 
     def _get_annotation_by_id(self, annotation_id: str) -> Sdmx30AnnotationModel | None:
@@ -95,110 +91,6 @@ class QuanthubSdmx21DataSet(UpdatedAtMixin, Sdmx21DataSet):
         if getter := mapping.get(property_source.source):
             return getter(property_source.field)
         raise ValueError(f"Unsupported property source: {property_source.source}")
-
-    def _get_data_explorer_url(
-        self,
-        base_url: str,
-        dsd: DataStructureDefinition,
-        key_dict: dict[str, list[str]],
-        sdmx_query: SdmxDataSetQuery,
-    ) -> str:
-        params = {
-            'urn': self._short_urn,
-            'filter': convert_keys_to_str(dsd, key_dict),
-        }
-
-        try:
-            if td_query := sdmx_query.time_dimension_query:
-                start_period = td_query.start_period
-                end_period = td_query.end_period
-            else:
-                start_period = None
-                end_period = None
-
-            if start_period:
-                params['startPeriod'] = start_period if '-' in start_period else f"{start_period}-A"
-            if end_period:
-                params['endPeriod'] = end_period if '-' in end_period else f"{end_period}-A"
-        except Exception as e:
-            _log.exception(e)
-
-        params_str = '&'.join([f'{k}={v}' for k, v in params.items()])
-        return f"{base_url}?{params_str}"
-
-    async def query(
-        self, query: DataSetQuery, auth_context: AuthContext
-    ) -> Sdmx21DataResponse | None:
-        status, missing_dimensions = self._evaluate_data_query_status(query)
-        if status != SdmxQueryReadinessStatus.READY:
-            raise ValueError(
-                f"Query is not ready: {query}, missing dimensions: {missing_dimensions}"
-            )
-        sdmx_query = self._to_sdmx_query(query)
-
-        try:
-            data_msg = await self._query_sdmx_data(sdmx_query, auth_context)
-        except Exception as e:
-            _log.exception(e)
-            return Sdmx21DataResponse(
-                dataset=self,
-                df=pd.DataFrame(),
-                sdmx_query=sdmx_query,
-                url=None,
-                status=DataResponseStatus(
-                    request_status=DataRequestStatus.FAILED,
-                    parsing_status=DataParsingStatus.NA,
-                ),
-            )
-
-        if not data_msg:
-            return Sdmx21DataResponse(
-                dataset=self,
-                df=pd.DataFrame(),
-                sdmx_query=sdmx_query,
-                url=None,
-                status=DataResponseStatus(
-                    request_status=DataRequestStatus.SUCCESS,
-                    parsing_status=DataParsingStatus.FAILED,
-                ),
-            )
-
-        if data_explorer_url := self._datasource.config.get_data_explorer_url():  # type: ignore
-            url = self._get_data_explorer_url(
-                base_url=data_explorer_url,
-                dsd=self._artefact.structure,
-                key_dict=sdmx_query.get_key(),
-                sdmx_query=sdmx_query,
-            )
-        else:
-            url = self._get_query_url(data_msg.response)  # type: ignore
-
-        try:
-            sdmx_pandas = await self._data_msg_to_dataframe(data_msg)
-            sdmx_pandas = await self._include_attributes(sdmx_pandas)
-        except Exception as e:
-            _log.exception(e)
-            return Sdmx21DataResponse(
-                dataset=self,
-                df=pd.DataFrame(),
-                sdmx_query=sdmx_query,
-                url=url,
-                status=DataResponseStatus(
-                    request_status=DataRequestStatus.SUCCESS,
-                    parsing_status=DataParsingStatus.FAILED,
-                ),
-            )
-
-        return Sdmx21DataResponse(
-            dataset=self,
-            df=sdmx_pandas,
-            sdmx_query=sdmx_query,
-            url=url,
-            status=DataResponseStatus(
-                request_status=DataRequestStatus.SUCCESS,
-                parsing_status=DataParsingStatus.SUCCESS,
-            ),
-        )
 
     def _availability_result_to_query(
         self, availability_result: StructureMessage
